+++ python/GafferSceneUI/ShuffleAttributesUI.py
@@ -1,6 +1,6 @@
 ##########################################################################
 #
-#  Copyright (c) 2015, Image Engine Design Inc. All rights reserved.
+#  Copyright (c) 2020, Image Engine Design Inc. All rights reserved.
 #
 #  Redistribution and use in source and binary forms, with or without
 #  modification, are permitted provided that the following conditions are
@@ -34,9 +34,39 @@
 #
 ##########################################################################
 
-from DocumentationTest import DocumentationTest
-from ArnoldShaderUITest import ArnoldShaderUITest
-from VisualiserAlgoTest import VisualiserAlgoTest
+import Gaffer
+import GafferUI
 
-if __name__ == "__main__":
-	unittest.main()
+import GafferScene
+
+Gaffer.Metadata.registerNode(
+
+	GafferScene.ShuffleAttributes,
+
+	"description",
+	"""
+	ShuffleAttributes is used to copy or rename arbitrary numbers of attributes at
+	the filtered locations. The deleteSource plugs may be used to remove the original
+	source attribute(s) after the shuffling has been completed.
+
+	An additional context variable \"${source}\" can be used on the destination plugs
+	to insert the name of each source attribute. For example, to prefix all attributes
+	with \"user:\" set the source to \"*\" and the destination to \"user:${source}\").
+	""",
+
+	plugs = {
+
+		"shuffles" : [
+
+			"description",
+			"""
+			The attributes to be shuffled - arbitrary numbers of attributes may be
+			shuffled via the source/destination plugs. The deleteSource plug may be
+			used to remove the original attribute(s).
+			""",
+
+		],
+
+	}
+
+)